fix volume in Alerts

+++ project/OsEngine/Alerts/AlertSignal.cs
@@ -23,7 +23,7 @@ namespace OsEngine.Alerts
         /// volume
         /// объём
         /// </summary>
-        public int Volume;
+        public decimal Volume;
 
         /// <summary>
         /// slippage
